% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haldensify.R
\name{cv_haldensify}
\alias{cv_haldensify}
\title{Conditional density estimation with HAL in a single cross-validation fold}
\usage{
cv_haldensify(fold, long_data, wts = rep(1, nrow(long_data)),
  lambda_seq = exp(seq(-1, -13, length = 1000)))
}
\arguments{
\item{fold}{Object specifying cross-validation folds as generated by a call
to \code{origami::make_folds}.}

\item{long_data}{A \code{data.table} or \code{data.frame} object containing
the data in long format, as given in \insertRef{diaz2011super}{haldensify},
as produced by \code{\link{format_long_hazards}}.}

\item{wts}{A \code{numeric} vector of observation-level weights, matching in
its length the number of records present in the long format data. The
default is to weight all observations equally.}

\item{lambda_seq}{A \code{numeric} sequence of values of the lambda tuning
parameter of the Lasso L1 regression, to be passed to \code{glmnet::glmnet}
through a call to \code{hal9001::fit_hal}.}
}
\description{
Conditional density estimation with HAL in a single cross-validation fold
}
