% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_bins}
\alias{make_bins}
\title{Histogram Binning Procedures for Pooled Hazards Regression}
\usage{
make_bins(grid_var, grid_type = c("hist", "scaled"), max_bins = 30L)
}
\arguments{
\item{grid_var}{The \code{numeric} vector over which histogram-based binning
is to be performed.}

\item{grid_type}{A \code{character} indicating the choice of binning rule,
with \code{"hist"} corresponding to the use of several rules proposed for
optimal histogram construction and \code{"scaled"} corresponding to the use
of various pre-set multiples of the square root of the sample size.}

\item{max_bins}{A \code{numeric} indicating the maximum number of bins that
are allowed in the grid for building the histogram based discretization.}
}
\description{
Histogram Binning Procedures for Pooled Hazards Regression
}
\keyword{internal}
