% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.haldensify}
\alias{plot.haldensify}
\title{Plot Method for HAL Conditional Density Estimates}
\usage{
\method{plot}{haldensify}(x, ..., type = c("risk", "density"))
}
\arguments{
\item{x}{[TO FILL IN]}

\item{...}{[TO FILL IN]}

\item{type}{[TO FILL IN]}
}
\value{
[TO FILL IN]
}
\description{
Plot Method for HAL Conditional Density Estimates
}
\examples{
# simulate data: W ~ U[-4, 4] and A|W ~ N(mu = W, sd = 0.5)
n_train <- 50
w <- runif(n_train, -4, 4)
a <- rnorm(n_train, w, 0.5)
# learn relationship A|W using HAL-based density estimation procedure
haldensify_fit <- haldensify(
  A = a, W = w, n_bins = 3,
  lambda_seq = exp(seq(-1, -10, length = 50))
)
plot(haldensify_fit, type = "risk")
}
