% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.haldensify}
\alias{predict.haldensify}
\title{Prediction method for HAL-based conditional density estimation}
\usage{
\method{predict}{haldensify}(object, ..., new_A, new_W)
}
\arguments{
\item{object}{An object of class \code{\link{haldensify}}, containing the
results of fitting the highly adaptive lasso for conditional density
estimation, as produced by a call to \code{\link{haldensify}}.}

\item{...}{Additional arguments passed to \code{predict} as necessary.}

\item{new_A}{The \code{numeric} vector or similar of the observed values of
an intervention for a group of observational units of interest.}

\item{new_W}{A \code{data.frame}, \code{matrix}, or similar giving the values
of baseline covariates (potential confounders) for the observed units whose
observed intervention values are provided in the previous argument.}
}
\description{
Prediction method for HAL-based conditional density estimation
}
