% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haldensify.R
\name{haldensify}
\alias{haldensify}
\title{Cross-validated conditional density estimation with HAL}
\usage{
haldensify(A, W, wts = rep(1, length(A)), grid_type = c("equal_range",
  "equal_mass", "equal_width"), n_bins = 10, width = NULL,
  lambda_seq = exp(seq(-1, -13, length = 1000)))
}
\arguments{
\item{A}{The \code{numeric} vector or similar of the observed values of an
intervention for a group of observational units of interest.}

\item{W}{A \code{data.frame}, \code{matrix}, or similar giving the values of
baseline covariates (potential confounders) for the observed units whose
observed intervention values are provided in the previous argument.}

\item{wts}{A \code{numeric} vector of observation-level weights. The default
is to weight all observations equally.}

\item{grid_type}{A \code{character} indicating the strategy to be used in
creating bins along the observed support of the intervention \code{A}. For
bins of equal range, use "equal_range", consulting the documentation of
\code{ggplot2::cut_interval} for more information. To ensure each bins has
the same number of points, use "equal_mass" and consult the documentation of
\code{ggplot2::cut_number} for details. If bins of equal width are desired,
use "equal_width" and consult \code{ggplot2::cut_width} as a reference.}

\item{n_bins}{Only used if \code{type} is set to \code{"equal_range"} or
\code{"equal_mass"}. This \code{numeric} value indicates the number of bins
that the support of the intervention \code{A} is to be divided into.}

\item{width}{Only used if \code{type} is set to \code{"equal_width"}. This
\code{numeric} provides the width of the bins to be produced along the
support of the observed values of the intervention \code{A}.}

\item{lambda_seq}{A \code{numeric} sequence of values of the lambda tuning
parameter of the Lasso L1 regression, to be passed to \code{glmnet::glmnet}
through a call to \code{hal9001::fit_hal}.}
}
\description{
Cross-validated conditional density estimation with HAL
}
