% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{map_hazard_to_density}
\alias{map_hazard_to_density}
\title{Map a predicted hazard to a predicted density for a single observation}
\usage{
map_hazard_to_density(hazard_pred_single_obs)
}
\arguments{
\item{hazard_pred_single_obs}{A \code{numeric} vector of the predicted hazard
of failure in a given bin (under a given partitioning of the support) for a
single observational unit based on a long format data structure (as produced
by \code{\link{format_long_hazards}}). This is simply the probability that
the observed value falls in a corresponding bin, given that it has not yet
failed (fallen in a previous bin), as given in
\insertRef{diaz2011super}{haldensify}.}
}
\description{
For a single observation, map a predicted hazard of failure (occurrence in a
particular bin, under a given partitioning of the support) to a density.
}
